requirements.txt
Django>=4.2.0
Pillow>=9.5.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1
python-dotenv>=1.0.0
whitenoise>=6.4.0
django-cleanup>=8.0.0
django-cors-headers>=4.3.1
django-extensions>=3.2.0
djangorestframework>=3.14.0
django-debug-toolbar>=4.1.0
django-allauth>=0.54.0
django-storages>=1.14.2
django-redis>=5.4.0
redis>=5.0.1
django-compressor>=4.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
celery>=5.3.6
django-ratelimit>=3.0.1
Stripe>=7.13.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
boto3>=1.34.14
django-filter>=23.5
django-taggit>=4.0.0
django-ckeditor>=6.7.0
django-ses>=3.5.2
django-countries>=7.5.1
django-phonenumber-field>=7.3.0
phonenumbers>=8.13.27
django-modeltranslation>=0.18.11
django-maintenance-mode>=0.19.0
django-import-export>=3.3.6
django-simple-history>=3.4.0
django-anymail>=10.2
sentry-sdk>=1.39.1
coverage>=7.2.5
pytest>=7.3.1
pytest-django>=4.5.2
pyOpenSSL>=23.2.0
